Add hgchangesets to backfill_derived_data
Summary:
This is currently a bit hacky, as bonsai->hg conversion does not work like other derived data, but it does the job.

We want to be able to dump Bonsai changesets from synced commits onto a bookmark, and have a tailer fill in Mercurial changesets behind us. This code does that via the existing mechanism for creating changesets

Reviewed By: StanislavGlebik

Differential Revision: D17552077

fbshipit-source-id: d5b13f884332e57e7d8c55256ccb4ea918fe4921