apiserver: use traits to convert to and from request types
Summary:
We will need to convert between the thrift request and response types and the
Mononoke API types in all of the functions.  Add traits to support this.

We can't use `std::convert::From` for this, as we need to
convert between types that are both defined in other crates.  Instead, define
`FromRequest` and `IntoResponse` which translate from request and to response
types accordingly.

Extend `thrift::CommitId` to provide useful methods on `thrift::CommitId`.

Reviewed By: krallin

Differential Revision: D17547875

fbshipit-source-id: 2f89802f5c4aa2df280bac699934abe23e94e4a6