mononoke_api: add files
Summary:
Add representations of files to the Mononoke API.

Like trees, files can exist either at paths within commits (which are
represented by a `ChangesetPathContext`, and may also refer to trees), or on
their own (which are represented by a `FileContext` and uniquely identified by
their content id, or one of their content hashes).  We keep these separate, as
history information is only relevant for `ChangesetPathContext`, and
`FileContext` may be more useful for clients that don't care about history.

Reviewed By: krallin

Differential Revision: D17547880

fbshipit-source-id: 4d2fdb27f4e9ebca98a9f7e902af064468a27e71