mononoke_api: add trees
Summary:
Add representations of trees to the Mononoke API.

Trees can exist either at paths within commits (which are represented by a
`ChangesetPathContext`, and may also refer to files), or on their own (which
are represented by a `TreeContext` and uniquely identified using their fsnode
id).  We keep these separate, as history information is only relevant for
`ChangesetPathContext`, and `TreeContext` may be more useful for clients that
don't care about history.

Reviewed By: krallin

Differential Revision: D17547881

fbshipit-source-id: be1b742303be84b816f632f1f1185f6fd1969d9e