sparse: warn when enabling non-existing profile
Summary:
Add a warning when a user enables a sparse profile on the command line that does not exist.

Triggering the check from multiple places, to not have to clutter the `_config` function. The `_config` function is called from other places as well (eg cloning a sparse repo) and I didn't want to add another guard parameter to it for such a benign functionality.

Besides the `_checknonexistingprofiles` function is only called in 3 situtations:

1. `hg sparse {enable|enableprofile}`
2. `hg sparse --enable-profile` (which is actually deprecated)
3. `hg clone --enable-profile`

Special care has been taken when cloning a sparse repo to do the check after the clone.

Reviewed By: markbt

Differential Revision: D17571442

fbshipit-source-id: 882dbe8b142bf192754dd17d0729d1bb0b472a83