mononoke: add find_first_synced query to synced_commit_mapping
Summary:
Given a list of changesets, a source and a target repo id, this query tells
you the first changeset from the list, which comes from the source repo and
has a synced equivalent in the target repo. Here, the word "first" refers to
the order of changesets in the provided list.

The naming is a little unfortunate, as the use case for this query is to
find the *last* (meaning newest) changeset that had been synced. But I wanted
the name to reflect this particular fn's behavior, rather than the intended
use.

Reviewed By: farnz

Differential Revision: D17600505

fbshipit-source-id: ea7336dcf839302e7b9ca6e1bc2ee88a3a1d99a9