Fall back to project name if path not present
Summary:
According to the [manifest specification][1], the `path` attribute of a
`<project>` is "An optional path ... If not supplied the project name is
used". Respect this when we parse manifests ourselves.

We may want to consider importing the manifest-parsing module from `repo`
itself and using that instead of rolling our own parsing code. That would
prevent bugs like this from happening at all.

[1] https://gerrit.googlesource.com/git-repo/+/master/docs/manifest-format.md

Reviewed By: mwdevine

Differential Revision: D17542715

fbshipit-source-id: 13d963f4d9b8bf30bcd5f5593aef0ad9da90e7ec