mononoke: change how force pushes are handled in the sync job
Summary:
There's a corner case that wasn't covered by the new sync job - if a user does
a "hg push --force" which pushes new commits to the server AND moves existing
bookmark then sync job just failed to sync. It tried to do a pushrebase of new
commits on top of existing bookmarks and that would result either in hash
mismatch or a pushrebase conflicts.

To fix it let's generate force pushrebase push in the sync job if bookmark did
non-fast forward move (i.e. "from" is not an ancestor of "to").

Note that there was an option to always do force pushrebase. That should work however that
would make stackpush optimization that we have on hg server impossible and that
might increase sync latency (see https://fburl.com/jpflbd67). Note that it's possible
that we'll switch to use normal push instead of pushrebase in the sync job,
however for now I'd like to go with pushrebase approach because it's exercised much more often.

Reviewed By: krallin

Differential Revision: D17624543

fbshipit-source-id: 1d6121a4767f67fbbbadae712ae8ab9d695857d5