add /{repo}/resolve_bookmark/{bookmark}
Summary:
scmquery_exporter is a tool to replace svn export and it needs 3 method from scmquery:
- log: to resolve a bookmark to a hash so later calls will be consistent
- ls and cat: to list and download file contents

ScmQuery proxy is currently not accessible from Corp to Prod VIP and will be available soon.
However people using the `js1 upgrade` tool is currently blocked due to the svn deprecation and we
have to make this tool work again ASAP.

The easiest way to make it work, comparing to all other worse approach, is to add this method
in mononoke api server and therefore we can add a backend in scmquery_export.

Discussion: https://fb.workplace.com/groups/JavaScriptOne/permalink/2454194331532724/

Reviewed By: StanislavGlebik

Differential Revision: D17605626

fbshipit-source-id: 19f229123cd980089e1e896c592a0f1afa3c531d