eden: gc now compacts sparseimage files for macos redirections
Summary:
The sparse images that we mount in place of bind mounts on macos
can grow over time, and don't automatically shrink if their contents are
purged.

This commit teaches `eden gc` to run `hdiutil compact` on these images
to reclaim space.

The output from `hdiutil compact` is really verbose and noisy, so we
suppress it unless there is an error.

Reviewed By: simpkins

Differential Revision: D17254235

fbshipit-source-id: 944b3824717c8b46d8170ec729c04a93323ab7f0