treemanifest: enable bfs diff by default
Summary: In order to support efficient data fetching, eventually BFS diff will need to be the default diff algorithm. As an initial step towards this, enable `treemanifest.bfsdiff` by default to validate that the BFS diff algorithm works as a drop-in replacement for the existing diff algorithm. Once we have confidence in this, we can land the next diff in this stack wherein we remove DFS diff entirely.

Reviewed By: xavierd

Differential Revision: D17618819

fbshipit-source-id: 444cb05874913e4eb267a24bd0596e6dde2e099e