always log thrift call times, even if the call fails
Summary:
Update the `Future` based version of the `ThriftLogHelper` class to log how
long the call took, even if it failed.

The code already logged on error for non-Future thrift handler functions, but
for `future_*()` calls we only logged when producing a successful result.
This switches the code from `thenValue()` to `thenTry()` so that we log on
both successful and failed results.

Reviewed By: wez

Differential Revision: D17636471

fbshipit-source-id: 859a326e5bc54183c80f16bf463ba6d451f624e5