testutil/dott: vendor Python 3's shlex library
Summary:
We need to get "whether an argument is single-quoted or not" information to
decide whether to run extra logic (ex. expanding globs) on it. `shlex` does not
really have public interface for that and it's hard to `wrapfunction` patch it.
Therefore vendor the script so we can customize it.

Some Python 2 compatibility work was done:
- use `cStringIO` instead of `io`
- mark the file as utf-8
- remove `quote`, which uses an incompatible `re.ASCII`

Reviewed By: xavierd

Differential Revision: D17675349

fbshipit-source-id: eb99a9b69ca15f770c410bd3845052cb0c6f1b17