edenfs: make `eden redirect add` nop if the config already exists
Summary:
mostafaeweda and otawfik observed that repeating the same `eden redirect
add` call with the same parameters was tearing down and re-mounting the bind
mount.  That shouldn't matter but it isn't as fast as not doing anything, and
sometimes hdiutil on mac can get into a weird state.

This commit compares the existing configuration against the new redirection
entry that we intend to add before applying the config; if they match then
it prints a note about not doing anything and then returns.

Reviewed By: mostafaeweda

Differential Revision: D17639592

fbshipit-source-id: 48d70d11e14a1a191ae655ad7a990c1c90c14953