manifest: walk manifest in breadth-first order
Summary:
Change the `Files` iterator in the Rust manifest code to traverse the tree in BFS order, allowing for layer-by-layer prefetching similar to `Diff`. This can substantially speed up walks over the tree when the cache is cold.

As a side-effect, this changes the order in which paths are reported during a manifest walk. (In particular, they are now reported in breadth-first order rather than depth-first order.) This may break things that rely on the existing ordering; as such, we may need to add a sort somewhere if this turns out to be a problem.

Reviewed By: xavierd

Differential Revision: D17645389

fbshipit-source-id: 624e426094a93e206bde4523ea8bd034fe5aeb90