snapshot: add the unionstore to handle the rebundling of snapshot bundles
Summary:
If some bundle consisted of several snapshots, and the client requests a part of that data, the server can trigger the rebundling process.
This diff teaches snapshots to the rebundling process.

The `bundlerepo` object should store data in memory, so a `unionstore` class is made up.
It allows to "pair" in-memory store with a regular local store and query them via the same API.

Reviewed By: markbt

Differential Revision: D17684945

fbshipit-source-id: 0bdb1f70b777fb2d8193b3bfa46b5915c2019663