eden: remove legacy bind mount support from doctor
Summary:
The doctor code can't see into the newer redirection
config, which makes it only partially valid.  In addition,
we run `eden redirect fixup` post-update; that supersedes
and replaces the doctor functionality.

I've also removed the associated tests: it's a shame that
they can't be re-purposed to work with the redirection code,
but their model is too awkward to use in that way--those tests
fake up aspects of the mount table and mount commands in layers
that are difficult to holistically fake up.

We have integration tests for the redirect functionality, so
we're not leaving things totally uncovered.

Reviewed By: simpkins

Differential Revision: D17252083

fbshipit-source-id: 8a576fb457bf0a4082bde1770085e809484ee73a