mononoke: fix parent order in blobimport
Summary:
Attached tasks have more details, but in short - if a commit has two parents
then the order of them in mercurial can change but the commit hash won't
change. In Mononoke it's different - change of parent order changes the hash.

This option allows us to fix parent order if needed

Reviewed By: krallin

Differential Revision: D17683807

fbshipit-source-id: 23549e55369b625ea5d597bc336b465373cbbc74