tests: convert test-globalrevs.t into -t.py test
Summary:
Converts the test-globalrevs.t test into the new format. Since the new
test format doesn't restart the process between invocations, we have to make the
globalrevs extension smarter about only executing its bits when the repository
in question actually has the extension enabled (instead of depending on the
process only living for the duration of a single command on a single repo).

Reviewed By: quark-zju

Differential Revision: D17693594

fbshipit-source-id: cae3c934af3c93ee2b5b83dcf82ba00b323dccac