match: fix tree.insert to mark short globs as unsure
Summary:
When inserting a glob into match._tree, it marks the tree as
'unsurerecursive' if the glob contains a complex component (i.e. a component
that contains "/"). If the glob was only one component, that check was not
executed against the component and therefore we could end up in a situation
where we didn't mark the tree as unsure. An example would be `'{*,{b,m}*/*}k'`,
which returned False for tree.visitdir("beans/black") when it should've returned
True.

Reviewed By: quark-zju

Differential Revision: D17118623

fbshipit-source-id: 50b6adbbecda13461c13c06a6e7d8523f8b29a2d