match: remove explicitdir
Summary:
match.explicitdir is used for commands to know which command line
patterns were explicit directories. It's existence makes implementing different
working copies more complicated because the matcher stops becoming a simple file
and directory tester and starts to become a more complicated traversal.

To simplify the upcoming working copy refactor, let's drop this functionality.
The only two places it's used (checking if we're commiting something inside a
provided directory during commit, and purge) can be replaced with just checking
if the provided path is actually a directory on disk.

Reviewed By: quark-zju

Differential Revision: D16951181

fbshipit-source-id: c03c79724e17db25f0490ab5906ef1aacfdf1634