treedirstate: don't filter when importing to treedirstate
Summary:
In a future diff I change the implementation of status, which broke
some tests and exposed a bug in treedirstate. Basically, when importing into
treedirstate it filtered the nonnormal files to only the ones that were tracked,
which meant it didn't include removed files. The new status implementation is
more sensitive to the nonnormal set being correct, so this broke.

The fix is to just not filter stuff when importing it into treedirstate. Nothing
uses treedirstate anymore, so this is probably a no-op for users.

Reviewed By: quark-zju

Differential Revision: D17079446

fbshipit-source-id: 8a41a792933472a3d0cec7732c4f546067a9470c