match: fix patternmatcher.visitdir
Summary:
patternmatcher can be used for non-exact and non-prefix patterns, but
the current visitdir implementation will return False unless a prefix pattern or
exact patterns are provided. A future change makes dirstate.status more reliant
on correct matcher behavior, and this breaks tests.

The fix is to return True (meaning that the directory should be visited)
whenever the pattern is not an explicit list of files. Hopefully we can move
this to our rust matcher in the future which will be able to optimize this case
around globs and regular expressions.

Reviewed By: quark-zju

Differential Revision: D17116369

fbshipit-source-id: d65b08060143efc3f4bbde3198b0c505a8ff4ec7