add: remove usage of dirstate.walk
Summary:
In a future diff we'll be restricting the dirstate interface. We'd like
to drop dirstate.walk, so let's remove it's use from `hg add`.

Note, the old code was actually a bit buggy. It had `full=False` which meant the
walk might not return clean files, which meant we might not warn about the user
attempting to add clean files, depending on the dirstate implementation. The new
code avoids that bug.

Reviewed By: quark-zju

Differential Revision: D17082285

fbshipit-source-id: 946f4edc852d8c62667fb50820f4b0c1c8f229a9