addremove: fix addremove in case insensitive situations
Summary:
My previous change, D17083282, broke addremove on certain case
insensitive filesystems. This was caused by me naively checking for file
existence, which didn't take into account case differences. Let's normalize the
file first and compare it. This appears to be what the old dirstate.walk code
does, though it's a huge mess so it's hard to be certain.

Reviewed By: singhsrb

Differential Revision: D17731308

fbshipit-source-id: 430edc1663748bff8db491bfc8ed609e2f340c60