indexedlog: migrate part of Index to new Error type (3)
Summary:
A lot of functions take (buf, checksum) tuple, instead of `Index` for input.
That is to avoid issues where borrowing the entire `Index` forbids modifying
other fields in `Index`.

However, not taking `Index` means it cannot figure out the file path on error.

To solve both problems, this diff defines a trait that is a subset of Index
including (on-disk buf, checksum, path). Then migrate functions from using
(buf, checksum) to the new trait (if it only needs to read from the on-disk
buffer), or &Index (if it also needs to work with in-memory dirty/mutable
data).

Reviewed By: xavierd

Differential Revision: D17705166

fbshipit-source-id: 90bde88142ea3718a2093beb02b8030d725a0e15