mononoke: use csid_resolve in places where we used resolve_hg_rev
Summary:
This is better from the usability perspective, as it can resolve hg changeset, bonsai changeset and a
bookmark.

However, this also means that when the user does provide an Hg commit hash and we need an hg commit hash, we are doing one more database query (`get_hg_from_bonsai_changeset`). This is a little bad, but all of the use-cases I changed here are from the admin tool, so we should be fine.

Reviewed By: StanislavGlebik

Differential Revision: D17740448

fbshipit-source-id: 7c1620979e25631f1a4e44d6310668fe634b2075