hgtime: new crate to work with time, esp for parsing
Summary:
I'd like a native `blackbox`-like command that accepts something like:

  --date 'since 3 hours ago'
  --date 'since yesterday'
  --date today
  --date 'Apr 2019' (implies a range of a month)
  --date 'Apr 2019 to Jul 2019'
  --date '13:00 to 15:00' (implies within today)

With a goal that this can be a superset and replace hg's date utilities,
and it can be used in `log --date`, revsets, etc.

Sadly, none of the top public Rust crates can sasity the existing features
("13:00" implying "today", "Apr 2019" implies a range of a month are two
less-known features that the Rust date utilties do not handle).

So I started porting the Python code (`mercurial.util.parsedate`) to Rust.

The core `parse` function is just 40 lines now.

Reviewed By: sfilipco

Differential Revision: D17504704

fbshipit-source-id: bde356a8cedf3298b060220de499e3954ed9d88b