Remove unused --sizes argument from eden debug file_stats command
Summary:
Since it has become cheap to fetch file sizes, we made a change in D17633671 to ignore the --sizes flag and always fetch sizes.

Now that we have removed the use of --sizes flag on the Sandcastle side in D17729705, we can remove the flag from the cli as well.

Reviewed By: fanzeyi

Differential Revision: D17746946

fbshipit-source-id: 1a943e32db6a6f6754a20d8e8fa3ac25f241f778