Add largest_directories_by_file_count to eden debug file_stats summary
Summary:
In order to optimize jobs using eden, we want to be able to see which directories we write/read a lot of files to/from. This information allows us to see if we can improve performance by adding certain directories to prefetch, or by symlinking directories that we write to often.

This diff adds to the summary a list of directories with at least 1000 files written to/read from them.

Reviewed By: fanzeyi

Differential Revision: D17747558

fbshipit-source-id: cf59beeee7c2eda9dea0dfa99f7ab92be9b33e1e