commitcloud: add commitcloud status to debugcrdump
Summary:
We eventually want to depend on all commits being uploaded to commit
cloud, instead of passing bundles to other services. To start this process,
let's make hg debugcrdump (which is used by jf submit) force the upload of the
given commits.

This required a minor refactor the backup.backup to make it more usable by logic
other than the backup command. I applied the same pattern to sync as well,
though it wasn't strictly necessary.

In a future diff we'll add logging to track how many jf submits send commits
that are not uploaded to commit cloud.

Reviewed By: mitrandir77

Differential Revision: D17721168

fbshipit-source-id: d191cfe3356a9bb2956a33921e13d317f7ed4ee5