clienttelemetry: log peer name to blackbox
Summary:
This updates clienttelemetry to log the peer name to the hg blackbox (on every
connection). This might make it easier to understand which host a given hg is
connecting to even if we can't get to its logs (e.g. in our external sync
processes, we are capturing output to match it for errors, so it's not trivial
to get the remote hostname even with clienttelemetry.announceremotehostname.

It's possible to retrieve the peer name with e.g.: ` hg blackbox --pattern '{"clienttelemetry": "_"}'`

Reviewed By: ikostia

Differential Revision: D17808834

fbshipit-source-id: 5c8afca6bd8b3bbbf14ec8d8efb3f9b1f6b25986