crdump: add branch to debugcrdump
Summary:
There are situations where it's useful for Phabricator to know what
branch a commit was created on. For instance, it may allow Phabricator to
suggest which branch to land a commit too.

Let's add the downstream branch to the crdump output, so jf can report it to
Phabricator.

Caveat: In Mercurial it's impossible to know for certain which remote bookmark a
local commit was made against. The best we can do is a heuristic.  The heuristic
is as follows:

1. For a given commit, find the first public ancestor.
2. For that ancestor commit, find all descendant commits that have remote
bookmarks.
3. If the descendant remote bookmarks contain "master", return that.
4. If the descendant remote bookmarks contain just one bookmark, return that.
5. Otherwise, return an empty string.

Reviewed By: mitrandir77

Differential Revision: D17814487

fbshipit-source-id: 1e37f57e18af55d4085fba27111b642a406c81af