revisionstore: handle same identical parents in topo_sort
Summary:
While importing the AOSP repo, we got into a situation where both parents were
the same node. This caused the assert at the end of the function to fire due to
some nodes not being added to the result.

As far as I can tell, this can happen, so let's deal with it.

Reviewed By: sfilipco

Differential Revision: D17832156

fbshipit-source-id: 9f968dcdc20aab4f4c00f6013fb1ee4501460159