remotefilelog: perform EdenAPI FFI calls in a new thread
Summary:
The Python interpreter is unable to respond to signals during FFI calls. This means that if there is a long-running FFI call (such as an HTTP tree prefetch via EdenAPI), the user will not be able to interrupt the process with Ctrl-C, leading to frustrating user experience.

To allow the interpreter to handle signals during FFI, let's wrap all EdenAPI FFI calls with a decorator that simply runs the operation in a separate Python thread and then busy-waits for the duration of the call (to avoid blocking in an uninterruptible way).

Reviewed By: quark-zju

Differential Revision: D17844316

fbshipit-source-id: baa4f489567105cfccdc314644ac1ae9e1c82a08