filesystem: fix issue with files in ignored directories
Summary:
Carolyn found an issue where if a file was in a directory that was
ignored, it would be reported as missing. This was because filesystem._walk
doesn't walk down ignored directories, and therefore did not report the file.
Let's fix it by having the remaining-files check look for the file on disk
instead of just assuming it is missing.

Reviewed By: xavierd

Differential Revision: D17848270

fbshipit-source-id: 7b6f0a38a224d76da2fa52bce155cdb47930b20a