edenapi: skip validation check for root treemanifest nodes
Summary:
Root tree manifests created in hybrid mode are identified by manifest node hashes that correspond to the equivalent flat manifest. As a result, the hash for such manifests do not match that of their content, which results in data validation failures in EdenAPI.

Given that we have these nodes are expected to exist in our repos due to the tree manifest migration, we need to handle this case. Let's do so by just skipping validation of these nodes and emitting a warning.

Reviewed By: quark-zju

Differential Revision: D17864271

fbshipit-source-id: b513c6f57e267e3c326191b23300e24bf7f92d77