manifest: add caching for trees in treemanifestlog
Summary:
Manifestlog is the object that is often passed around in mercurial as the proxy
to manfiests. Commits that need to check various information on manifests will
will recreate trees for every file that goes into the commit. This is a lot of
churn. Caching the manfiests in the manfiestlog is an easy fix right now.

This codepath was hitting the Rust manifest code path more than the C++ code
path because Rust will eagerly parse entries. Among parsing the highest cost
is to convert the hex Node to a binary Node - an operation that would not
be performed by C++.

Reviewed By: xavierd

Differential Revision: D17859964

fbshipit-source-id: 44341c1cd4e9b70d80d864065f234cb906a8d76b