mononoke: move PostResolverAction-related things from bundle2_resolver
Summary:
See bottom diff of the stack for the entire stack's purpose.

The goal of this particular diff is to address step 2:

Things that are only used by `repo_client::unbundle` and not by `bundle2_resolver`, should be moved out of `bundle2_resolver`. While moving such things, it will turn out that some things that `Bundle2Resolver` struct owned (and some things that `bundle2_resolver::resolver::resolve` function accepted as arguments) are only threaded down to whichever post-resolve action is appropriate. Such things should just not be passed into `resolve` in the first place.

In theory, after this diff it is already possible to introduce something in between the `resolve` and `run_post_resolve_action` calls.

Reviewed By: StanislavGlebik

Differential Revision: D17853603

fbshipit-source-id: 0442b478dad82723383896666096eec2aec64e58