backfill_derived_data: allow access to the unredacted repo
Summary:
Some backfill types, e.g. fsnodes, require access to the unredacted repo in
order to correctly compute their values.  For example, fsnodes require the
content hashes of the redacted blobs to correctly form the fnsode graph.
This kind of access is ok, as we are only deriving summary information
(hashes), and won't reveal the content of the redacted blob.

Add the ability for `backfill_derive_data` to access the unredacted repo,
limited only to whitelisted derived data types.

Reviewed By: ikostia

Differential Revision: D17877191

fbshipit-source-id: 6c9b1dfdfb2e6f431815ddf3de60029fbb180454