filesystem: only set dirstate to dirty if a file was actually marked clean
Summary:
A recent dirstate refactor made the dirstate get marked dirty if a file
was marked for clean. Unfortunately it also set the dirstate to dirty if the
file was only checked for clean and didn't require marking. This resulted in the
dirstate getting written with no changes, which triggered IDE's to run hg
status, which re-wrote the dirstate, triggered IDE's, etc.

The fix is to use the built in dirstate.clearneedcheck function, which handles
this appropriately.

Reviewed By: xavierd

Differential Revision: D17919381

fbshipit-source-id: 78698f1c82f00a95502d09ad9515bee58b1191fc