dirstate: fix hg diff crashing for deleted adds
Summary:
The new dirstate loops over the nonnormal files and marks them as
added, removed, etc. There was a bug where if a file was added, it was marked
added even if it disappeared from disk. Let's double check that the file exists
before we mark it as added.

I don't believe the other file states (merged and removed mainly) have this issue.

Normally this issue only reproduced in Eden checkouts. This highlighted a bug
where the non-eden filesystem was reporting files that had been
hg-added-then-rm'd as changed, when really pendingchanges should not report that
since that file being missing it not a pending change relative to the originally
checked out tree.

Reviewed By: quark-zju

Differential Revision: D17921401

fbshipit-source-id: a57992b83173c299ba3307d6324dd357cd2b0b24