mononoke: make RepoClient struct own a commit sync config
Summary:
This is part of work on the push redirector, which works in this scenario:
1. an engineer tries to push a commit to `fbsource`/`ovrsource` (aka small repo)
1. push redirector makes sure that this push is redirected to `megarepo_test` (aka large repo). This means:
   1. saving every uploaded changeset from the bundle2 into the original repo
   1. syncing all of the uploaded changesets in the target (large) repo
   1. parsing the instructions of what has to be done with the uploaded stuff
   1. transforming these instructions so that they apply to the target repo
   1. running these instructions in the target repo, noticing which changesets and bookmarks they produce
   1. syncing these produced changesets back in to the original (small) repo

Steps 2.b, 2.d, 2.e and 2.f require `RepoClient` to know about the target repo and to know how to sync commits back and forth and how to change bookmark names.

Current `RepoSyncTarget` is basically my best guess of what may be needed, but the important bit is that it's threaded through and to add/remove something it is enough to change some `repo_handlers` code and that's it.

Reviewed By: StanislavGlebik

Differential Revision: D17900899

fbshipit-source-id: 5bc2b5dac5ecc7dcef855b9c8f91fa8a618fbbad