mononoke/hg_sync_job: verify LFS blob presence when generating bundles
Summary:
When we generate bundles, we choose which files to send as LFS to the remote. Right now, we're configured so that the Mononoke and client thresholds match, and our LFS server is designed to replicate blobs into Dewey synchronously, so this should be safe.

That said, to further minimize the risk of breakage, we can verify that the blobs we are about to send as LFS are indeed present in the other backend's LFS server. This does that.

Reviewed By: StanislavGlebik

Differential Revision: D17957232

fbshipit-source-id: bfc77470377afcc94537fb627c214a94e8f08c0b