mononoke: separate push vs infinitepush in unbundle handling
Summary:
Currently, it is possible to combine normal, pure push and infinitepush push in the same `unbundle` wireproto command. I would argue that this is a bad thing, as they are pretty different beasts. This has lead to a certain amount of functions that have to deal with multiple variants of enums and, IMO, has made code harder to understand. This diff tries to separate these concerns from the perspective of `PostResolveAction` and below: separate all the verification, conversion and processing functions down until we hit the database layer.

It also gets rid of "resolve" naming in one place, as it has no clear semantic meaning (compared to "save to db" at least) IMO.

Note that this is not just a tech debt thing: for push redirector we need to know which scenario we are in to apply the correct redirection semantics.

Reviewed By: farnz

Differential Revision: D17966069

fbshipit-source-id: e7fe8384a35a0f5945e89b91166e11f2f675e79f